build: completed fix. WDS_SOCKET issue. (See details)
finally fixed by realizing it was a WDS_SOCKET issue. This happens with CRA when trying to map to a new port even through the dev-server was defauling to its 3000 port. To get around this, I realized it had nothing to do with the backend api -- except for the nodemon -- and I changed backend back to how it was, then just mapped the capstone ui to port 3000 instead of 3001 to give it what it wanted. This fixes the docker compose up issue. There is another work around, but it requesties setting an environmental variable WDS_SOCKET_PORT=0